Move `start_server` to library
And use it in integration tests.
