+++ backend/src/bin/abacus.rs
@@ -1,7 +1,5 @@
 #[cfg(feature = "dev-database")]
-use abacus::fixtures;
-use abacus::router;
-use axum::serve::ListenerExt;
+use abacus::{fixtures, start_server};
 use clap::Parser;
 use sqlx::{SqlitePool, sqlite::SqliteConnectOptions};
 use std::{
@@ -9,8 +7,8 @@ use std::{
     net::{Ipv4Addr, SocketAddr},
     str::FromStr,
 };
-use tokio::{net::TcpListener, signal};
-use tracing::{info, level_filters::LevelFilter, trace};
+use tokio::net::TcpListener;
+use tracing::{info, level_filters::LevelFilter};
 use tracing_subscriber::EnvFilter;
 
 /// Abacus API and asset server
@@ -33,27 +31,6 @@ struct Args {
     #[cfg(feature = "dev-database")]
     #[arg(short, long)]
     reset_database: bool,
-}
-
-/// Start the API server on the given port, using the given database pool.
-async fn start_server(pool: SqlitePool, listener: TcpListener) -> Result<(), Box<dyn Error>> {
-    let app = router(pool)?;
-
-    info!("Starting API server on http://{}", listener.local_addr()?);
-    let listener = listener.tap_io(|tcp_stream| {
-        if let Err(err) = tcp_stream.set_nodelay(true) {
-            trace!("failed to set TCP_NODELAY on incoming connection: {err:#}");
-        }
-    });
-
-    axum::serve(
-        listener,
-        app.into_make_service_with_connect_info::<SocketAddr>(),
-    )
-    .with_graceful_shutdown(shutdown_signal())
-    .await?;
-
-    Ok(())
 }
 
 /// Main entry point for the application. Sets up the database, and starts the
@@ -101,61 +78,4 @@ async fn create_sqlite_pool(
     }
 
     Ok(pool)
-}
-
-/// Graceful shutdown, useful for Docker containers.
-///
-/// Copied from the
-/// [axum graceful-shutdown example](https://github.com/tokio-rs/axum/blob/6318b57fda6b524b4d3c7909e07946e2b246ebd2/examples/graceful-shutdown/src/main.rs)
-/// (under the MIT license).
-async fn shutdown_signal() {
-    let ctrl_c = async {
-        signal::ctrl_c()
-            .await
-            .expect("failed to install Ctrl+C handler");
-    };
-
-    #[cfg(unix)]
-    let terminate = async {
-        signal::unix::signal(signal::unix::SignalKind::terminate())
-            .expect("failed to install signal handler")
-            .recv()
-            .await;
-    };
-
-    #[cfg(not(unix))]
-    let terminate = std::future::pending::<()>();
-
-    tokio::select! {
-        _ = ctrl_c => {},
-        _ = terminate => {},
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use sqlx::SqlitePool;
-    use test_log::test;
-    use tokio::net::TcpListener;
-
-    use super::start_server;
-
-    #[test(sqlx::test)]
-    async fn test_abacus_starts(pool: SqlitePool) {
-        let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-
-        let task = tokio::spawn(async move {
-            start_server(pool, listener).await.unwrap();
-        });
-
-        let result = reqwest::get(format!("http://{addr}/api/user/whoami"))
-            .await
-            .unwrap();
-
-        assert_eq!(result.status(), 401);
-
-        task.abort();
-        let _ = task.await;
-    }
 }
